chore: all SIO services are in a services/sio/
