+++ src/services/sio/pumps-sio.service.ts
@@ -1,9 +1,9 @@
 import { Subscriber } from "rxjs";
 
-import { PumpHistorical } from "../models/entities/pump-historical";
+import { PumpHistorical } from "../../models/entities/pump-historical";
 
-import { pumpHistoricalRepository } from "../models/repositories";
-import { SIOService } from "./sio.service";
+import { pumpHistoricalRepository } from "../../models/repositories";
+import { SIOService } from "../sio.service";
 
 const enum NamespaceNames {
   LastPumpHistoricals = "/pumps/last-historicals"
